+++ src/V1/guards/index.ts
@@ -33,5 +33,4 @@
 //
 
 export * from "./isUuidData";
-export * from "./isUuidString";
 export * from "./isUuidType";