+++ src/V1/guards/isUuidData.ts
@@ -30,16 +30,16 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { isUuidString, Uuid } from "..";
+
+/**
+ * A regex that will match UUID v1-v5, and the NULL UUID
+ */
+export const UuidRegex = new RegExp("^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$", "i");
 
 /**
  * returns `true` if the given input is a UUID that we can work with,
  * `false` otherwise
  */
-export function isUuidData(input: Uuid|string): boolean {
-    if (typeof(input) === "string") {
-        return isUuidString(input);
-    }
-
-    return true;
+export function isUuidData(input: string): boolean {
+    return UuidRegex.test(input);
 }