+++ src/V1/transforms/bytes.spec.ts
@@ -31,14 +31,14 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { Uuid } from "../";
-import { UuidByteLength } from "../types";
+import { UuidByteLength } from "../types/Uuid";
 import { uuidFromBytes, uuidToBytes } from "./bytes";
+import { uuidFromFormatted } from "./formatted";
 
 describe("uuidToBytes()", () => {
 
     it("accepts a UUID object", () => {
-        const inputValue = new Uuid("123e4567-e89b-12d3-a456-426655440000");
+        const inputValue = uuidFromFormatted("123e4567-e89b-12d3-a456-426655440000");
         const expectedValue = Buffer.from("123e4567e89b12d3a456426655440000", "hex");
 
         const actualValue = uuidToBytes(inputValue);
@@ -47,7 +47,7 @@ describe("uuidToBytes()", () => {
     });
 
     it("accepts a Buffer to write to", () => {
-        const inputValue = new Uuid("123e4567-e89b-12d3-a456-426655440000");
+        const inputValue = uuidFromFormatted("123e4567-e89b-12d3-a456-426655440000");
         const inputBuffer = Buffer.alloc(UuidByteLength);
         const expectedValue = Buffer.from("123e4567e89b12d3a456426655440000", "hex");
 
@@ -62,7 +62,7 @@ describe("uuidToBytes()", () => {
 describe("uuidFromBytes()", () => {
 
     it("accepts an array of bytes", () => {
-        const expectedValue = new Uuid("123e4567-e89b-12d3-a456-426655440000");
+        const expectedValue = uuidFromFormatted("123e4567-e89b-12d3-a456-426655440000");
         const inputValue = Buffer.from("123e4567e89b12d3a456426655440000", "hex");
 
         const actualValue = uuidFromBytes(inputValue);